Add/Use ApplianceConsoleCli for validation
Updates tests to now use the newly added `appliance_console_cli`
commands for db restore instead of `rake` from `vmdb`.

This has a couple of changes to support this:

- Updates the appliance_sync.sh to be alphabetical and adds a few more
  files to the sync
- Updates the appliance_rake_restore_patches.rb to be more genericly
  named and support the `appliance_console_cli` version of
  `PostgresAdmin` better
- Have the ApplianceSecondaryDB use the `postgres` user instead of
  `vagrant` (less changes required to patch `PostgresAdmin` with)
- Update the `DbValidator` to use the `ApplianceConsoleCli` helper
  library instead of `RakeHelper` (mostly the former is a slight
  re-write of the latter)
