+++ test/support/appliance_restore_patches.rb
@@ -6,7 +6,7 @@ require "linux_admin"
 # is loaded.
 module LinuxAdmin
   def Service.new(service_name)
-    if PostgresAdmin.service_name == service_name
+    if ManageIQ::ApplianceConsole::PostgresAdmin.service_name == service_name
       ApplianceSecondaryDB
     else
       super
@@ -14,14 +14,8 @@ module LinuxAdmin
   end
 end
 
-require "postgres_admin"
-
-# Patch PostgresAdmin to use different user/group
-class PostgresAdmin
-  def self.user
-    "vagrant".freeze
-  end
-end
+console_codebase_dir = Dir["/opt/manageiq/manageiq-gemset/gems/manageiq-appliance_console-*"].first
+require "#{console_codebase_dir}/lib/manageiq/appliance_console/postgres_admin.rb"
 
 ENV["APPLIANCE_PG_DATA"]    = "/opt/manageiq/postgres_restore_pg/data"
 ENV["APPLIANCE_PG_SERVICE"] = "local_pg_instance"
